% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{alert_on_fails}
\alias{alert_on_fails}
\title{Generates a cli alert if any requests have failed}
\usage{
alert_on_fails(resps)
}
\arguments{
\item{resps}{A list of API responses}
}
\value{
The list of resps that was passed in
}
\description{
Generates a cli alert if any requests have failed
}
\keyword{internal}
