% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{convert_roles}
\alias{convert_roles}
\title{Converts a vector of role labels such as "trust" to a single string of codes}
\usage{
convert_roles(roles)
}
\arguments{
\item{roles}{A character vector}
}
\value{
A character scalar (string)
}
\description{
Converts a vector of role labels such as "trust" to a single string of codes
}
\keyword{internal}
