% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{core_request}
\alias{core_request}
\title{A core httr2 \code{request} to be used by wrapper functions}
\usage{
core_request(api_url = "https://sandbox.api.service.nhs.uk")
}
\arguments{
\item{api_url}{The base URL of the NHS England Organisation Data Terminology
FHIR API Service}
}
\description{
A core httr2 \code{request} to be used by wrapper functions
}
\keyword{internal}
