% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{extract_resource_data}
\alias{extract_resource_data}
\title{Extracts data from a list of API responses}
\usage{
extract_resource_data(resps)
}
\description{
Extracts data from a list of API responses
}
\keyword{internal}
