% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_organisation_details}
\alias{get_organisation_details}
\title{Query the Organization API using an organisation code}
\usage{
get_organisation_details(org_code, ...)
}
\arguments{
\item{org_code}{string - the organisation code}

\item{...}{Used for supplying an alternative API URL. Use only if needed.}
}
\value{
All data as a list
}
\description{
See https://digital.nhs.uk/developer/api-catalogue/organisation-data-terminology#get-/Organization/-id-
}
