% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_operated_by}
\alias{is_operated_by}
\title{A test to see if a relationship value is equal to "RE6" (IS_OPERATED_BY)}
\usage{
is_operated_by(structure)
}
\arguments{
\item{structure}{A section of a list}
}
\value{
logical TRUE or FALSE
}
\description{
A test to see if a relationship value is equal to "RE6" (IS_OPERATED_BY)
}
\keyword{internal}
