% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{org_affiliation_request}
\alias{org_affiliation_request}
\title{Basic request to OrganizationAffiliation API}
\usage{
org_affiliation_request(...)
}
\arguments{
\item{...}{Used for supplying an alternative API URL. Use only if needed.}
}
\description{
Basic request to OrganizationAffiliation API
}
\keyword{internal}
