% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{org_req}
\alias{org_req}
\title{Simple httr2 request passing the og code as an ID to the Organization API}
\usage{
org_req(id, ...)
}
\arguments{
\item{id}{string An NHS England organisation code}

\item{...}{Used for supplying an alternative API URL. Use only if needed.}
}
\description{
Simple httr2 request passing the og code as an ID to the Organization API
}
\keyword{internal}
