% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{organisation_request}
\alias{organisation_request}
\title{Basic request to the Organization API}
\usage{
organisation_request(...)
}
\arguments{
\item{...}{Used for supplying an alternative API URL. Use only if needed.}
}
\description{
Basic request to the Organization API
}
\keyword{internal}
