% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{poss_map}
\alias{poss_map}
\title{A "safe" version of purrr::map_chr that returns NA if a value is not found}
\usage{
poss_map(...)
}
\description{
A "safe" version of purrr::map_chr that returns NA if a value is not found
}
\keyword{internal}
