% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{get_name_from_org_code}
\alias{get_name_from_org_code}
\title{Return an organisation name, based on its organisation code}
\usage{
get_name_from_org_code(org_code, ...)
}
\arguments{
\item{org_code}{string An NHS England organisation code}

\item{...}{Used for supplying an alternative API URL. Use only if needed.}
}
\value{
A named character vector of length 1. The organisation name is the
content, with the org code as name of the vector
}
\description{
Return an organisation name, based on its organisation code
}
