% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_next_link_url}
\alias{get_next_link_url}
\title{A helper function to extract the next link from a response and return an
amended request, or NULL if all results have now been returned}
\usage{
get_next_link_url(resp, req)
}
\description{
A helper function to extract the next link from a response and return an
amended request, or NULL if all results have now been returned
}
\keyword{internal}
