% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_org_info}
\alias{get_org_info}
\title{Return information about organisations from a name search}
\usage{
get_org_info(org_name, type = c("starts", "contains", "exact"))
}
\arguments{
\item{org_name}{string Part or whole organisation name to search for}

\item{type}{string Type of name search to perform}
}
\value{
A tibble with 2 columns: \code{org_code} and \code{org_name}, and a row for
each organisation matched by \code{name} (according to the chosen search \code{type})
}
\description{
The name search can be partial or exact. There are three flavours of search
available:
\itemize{
\item "starts": search for an organisation name that starts with \code{name}
\item "contains": search for an organisation name that contains \code{name} anywhere
\item "exact": search for an organisation name that exactly matches \code{name}
Use the \code{type} argument to specify which one you want. \code{"starts"} is the
default option.
}
}
