% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sites_from_org_code}
\alias{get_sites_from_org_code}
\title{Return affiliated sites of an NHS Trust, based on the Trust organisation code}
\usage{
get_sites_from_org_code(org_code, ...)
}
\arguments{
\item{org_code}{string - the organisation code}

\item{...}{Used for supplying an alternative API URL. Use only if needed.}
}
\value{
A character vector of site organisation codes
}
\description{
Return affiliated sites of an NHS Trust, based on the Trust organisation code
}
\keyword{internal}
