% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{is_trust}
\alias{is_trust}
\title{Check to see if a code represents an organisation that is an NHS Trust}
\usage{
is_trust(org_code, ...)
}
\value{
logical TRUE or FALSE
}
\description{
Check to see if a code represents an organisation that is an NHS Trust
}
\keyword{internal}
