% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_trust_site}
\alias{is_trust_site}
\title{Check to see if code represent NHS Trust sites}
\usage{
is_trust_site(org_codes, ...)
}
\arguments{
\item{org_codes}{A character vector of organisation codes}
}
\value{
A logical vector of TRUE or FALSE values
}
\description{
Check to see if code represent NHS Trust sites
}
\keyword{internal}
