% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odsapique-package.R
\docType{package}
\name{odsapique-package}
\alias{odsapique}
\alias{odsapique-package}
\title{odsapique: Query the NHS Organisation Data Terminology FHIR API}
\description{
Search for an organisation name or details.
}
\author{
\strong{Maintainer}: Fran Barton \email{francis.barton@nhs.net} (\href{https://orcid.org/0000-0002-5650-1176}{ORCID})

Authors:
\itemize{
  \item Fran Barton \email{francis.barton@nhs.net} (\href{https://orcid.org/0000-0002-5650-1176}{ORCID})
}

}
\keyword{internal}
