% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{org_role_location}
\alias{org_role_location}
\title{Returns a list of elements to use as a location for the poss_map accessor fn.}
\usage{
org_role_location(which = c("code", "display"))
}
\arguments{
\item{which}{string Whether to access the organisation role code or the
organisation role name ("display")}
}
\value{
A list
}
\description{
Returns a list of elements to use as a location for the poss_map accessor fn.
}
\keyword{internal}
